backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pyinstaller>=6.18.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
